+++ src/detection/locale/locale_linux.c
@@ -8,19 +8,9 @@ void ffDetectLocale(FFstrbuf* result)
     if(result->length > 0)
         return;
 
-    ffStrbufAppendS(result, getenv("LC_MESSAGES"));
-    if(result->length > 0)
-        return;
-
     ffStrbufAppendS(result, getenv("LANG"));
     if(result->length > 0)
         return;
-
-    #ifdef LC_MESSAGES
-    ffStrbufAppendS(result, setlocale(LC_MESSAGES, NULL));
-    if(result->length > 0)
-        return;
-    #endif
 
     ffStrbufAppendS(result, setlocale(LC_TIME, NULL));
 }