+++ src/util/binary.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "fastfetch.h"
+
+const char* ffBinaryExtractStrings(const char* file, bool (*cb)(const char* str, uint32_t len, void* userdata), void* userdata);